move transport stream hooks to webapp fetchers
