+++ apps/webapp/src/fetchers/assets.ts
@@ -1,4 +1,4 @@
-import { streamToPromise } from '@penumbra-zone/transport';
+import { streamToPromise } from './stream.ts';
 import { AssetsRequest } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/view/v1alpha1/view_pb';
 import { viewClient } from '../clients/grpc.ts';
