+++ apps/webapp/src/fetchers/stream.ts
