+++ packages/transport/src/index.ts
@@ -2,6 +2,5 @@ export * from './create';
 export * from './types';
 export * from './unary';
 export * from './stream';
-export * from './hooks';
 export * from './looper';
 export * from './std-client';
